rename to car registration field
