+++ apps/store/src/components/PriceForm/CarRegistrationField.tsx
